Consuming Absam's API (BETA)
Create two functions the first is designed for suspend a server, the second one is for unsuspend. In absam if you place a server in suspension mode you only need to pay a third of the price. That means if your server costs 100 BRL in suspension mode you'll need to pay 30 BRL.